Introduced AudioEngine as an EnvironmentObject:  class to centralise all audio related logic
