+++ Hamba/View/MapView/MapView.swift
@@ -10,6 +10,7 @@ import MapKit
 import SwiftUI
 
 struct MapView: View {
+    @EnvironmentObject var audioEngine: AudioEngine
     @StateObject var mapViewModel: MapViewModel
 
     var body: some View {
@@ -22,7 +23,7 @@ struct MapView: View {
         }
         .phoneOnlyNavigationView()
         .onAppear(perform: {
-            playSound(sound: "focus-loop-corporate-music-114297", type: "mp3")
+            audioEngine.playSound(sound: "focus-loop-corporate-music-114297", type: "mp3")
             DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
                 mapViewModel.checkIfLocationServicesIsEnabled()
             }
@@ -32,11 +33,6 @@ struct MapView: View {
     var navBar: some View {
         VStack {
             mainNavBar(mapViewModel: mapViewModel)
-                .background(
-                    // Apply a linear gradient with a blur effect
-                    LinearGradient(gradient: Gradient(colors: [Color.clear]), startPoint: .top, endPoint: .bottom)
-                        .blur(radius: 0) // Adjust the blur radius as needed
-                )
             Spacer()
         }
     }
